+++ tools/test_runner.go
@@ -1,6 +1,3 @@
-// Package main provides a comprehensive test runner and performance analyzer for zod-go
-// This tool helps developers understand the performance characteristics and capabilities
-// of the enhanced DX API compared to the legacy implementation.
 package main
 
 import (
@@ -15,9 +12,9 @@ import (
 )
 
 func main() {
-	fmt.Println("🧪 zod-go Enhanced DX Test Runner & Performance Analyzer")
+	fmt.Println("zod-go Enhanced DX Test Runner & Performance Analyzer")
 	fmt.Println("=====================================================")
-	
+
 	runPerformanceAnalysis()
 	runMemoryAnalysis()
 	runConcurrencyTests()
@@ -26,9 +23,8 @@ func main() {
 	generatePerformanceReport()
 }
 
-// Performance analysis for different API patterns
 func runPerformanceAnalysis() {
-	fmt.Println("\n📊 Performance Analysis")
+	fmt.Println("\nPerformance Analysis")
 	fmt.Println("======================")
 
 	scenarios := []struct {
@@ -110,21 +106,19 @@ func runPerformanceAnalysis() {
 
 	for _, scenario := range scenarios {
 		fmt.Printf("\n🔍 Testing: %s - %s\n", scenario.name, scenario.description)
-		
-		// Warm-up
+
 		schema := scenario.setupFunc()
 		for i := 0; i < 100; i++ {
 			_ = schema.Validate(scenario.testData)
 		}
 
-		// Benchmark
 		iterations := 10000
 		start := time.Now()
-		
+
 		for i := 0; i < iterations; i++ {
 			_ = schema.Validate(scenario.testData)
 		}
-		
+
 		duration := time.Since(start)
 		avgTime := duration / time.Duration(iterations)
 		opsPerSec := float64(iterations) / duration.Seconds()
@@ -132,35 +126,32 @@ func runPerformanceAnalysis() {
 		fmt.Printf("   ⏱️  Total time: %v\n", duration)
 		fmt.Printf("   📈 Average per operation: %v\n", avgTime)
 		fmt.Printf("   🚀 Operations per second: %.0f\n", opsPerSec)
-		
-		// Memory usage during validation
+
 		var m1, m2 runtime.MemStats
 		runtime.GC()
 		runtime.ReadMemStats(&m1)
-		
+
 		for i := 0; i < 1000; i++ {
 			_ = schema.Validate(scenario.testData)
 		}
-		
+
 		runtime.ReadMemStats(&m2)
 		memUsed := m2.TotalAlloc - m1.TotalAlloc
 		fmt.Printf("   💾 Memory per 1K ops: %d bytes\n", memUsed)
 	}
 }
 
-// Memory usage analysis
 func runMemoryAnalysis() {
 	fmt.Println("\n💾 Memory Analysis")
 	fmt.Println("==================")
 
 	fmt.Println("\n📋 Schema Creation Memory Usage:")
-	
+
 	var m1, m2 runtime.MemStats
-	
-	// Test schema creation memory usage
+
 	runtime.GC()
 	runtime.ReadMemStats(&m1)
-	
+
 	schemas := make([]interface{ Validate(interface{}) error }, 1000)
 	for i := 0; i < 1000; i++ {
 		schemas[i] = validators.String().
@@ -170,26 +161,25 @@ func runMemoryAnalysis() {
 			Required().
 			WithMessage(validators.Errors.MinLength(), "Custom message")
 	}
-	
+
 	runtime.ReadMemStats(&m2)
 	schemaMemory := m2.TotalAlloc - m1.TotalAlloc
-	fmt.Printf("   🏗️  1000 string schemas: %d bytes (avg: %d bytes/schema)\n", 
+	fmt.Printf("   🏗️  1000 string schemas: %d bytes (avg: %d bytes/schema)\n",
 		schemaMemory, schemaMemory/1000)
 
-	// Test complex object schema memory
 	runtime.GC()
 	runtime.ReadMemStats(&m1)
-	
+
 	complexSchemas := make([]interface{ Validate(interface{}) error }, 100)
 	for i := 0; i < 100; i++ {
 		complexSchemas[i] = validators.Object(map[string]interface{}{
 			"user": validators.Object(map[string]interface{}{
-				"name":    validators.String().Min(1).Required(),
-				"email":   validators.Email(),
-				"age":     validators.Number().Integer().Optional(),
+				"name":  validators.String().Min(1).Required(),
+				"email": validators.Email(),
+				"age":   validators.Number().Integer().Optional(),
 				"profile": validators.Object(map[string]interface{}{
 					"bio":     validators.String().Max(500).Optional(),
-					"website": validators.URL().Optional(),
+					"website": validators.String().URL().Optional(),
 				}).Optional(),
 			}).Required(),
 			"settings": validators.Object(map[string]interface{}{
@@ -198,13 +188,12 @@ func runMemoryAnalysis() {
 			}).Optional(),
 		}).Required()
 	}
-	
+
 	runtime.ReadMemStats(&m2)
 	complexMemory := m2.TotalAlloc - m1.TotalAlloc
-	fmt.Printf("   🏗️  100 complex schemas: %d bytes (avg: %d bytes/schema)\n", 
+	fmt.Printf("   🏗️  100 complex schemas: %d bytes (avg: %d bytes/schema)\n",
 		complexMemory, complexMemory/100)
 
-	// Test validation memory usage
 	schema := validators.Object(map[string]interface{}{
 		"name":  validators.String().Required(),
 		"items": validators.Array(validators.String()).Required(),
@@ -212,7 +201,7 @@ func runMemoryAnalysis() {
 
 	runtime.GC()
 	runtime.ReadMemStats(&m1)
-	
+
 	for i := 0; i < 1000; i++ {
 		testData := map[string]interface{}{
 			"name":  fmt.Sprintf("test_%d", i),
@@ -220,16 +209,15 @@ func runMemoryAnalysis() {
 		}
 		_ = schema.Validate(testData)
 	}
-	
+
 	runtime.ReadMemStats(&m2)
 	validationMemory := m2.TotalAlloc - m1.TotalAlloc
-	fmt.Printf("   ✅ 1000 validations: %d bytes (avg: %d bytes/validation)\n", 
+	fmt.Printf("   ✅ 1000 validations: %d bytes (avg: %d bytes/validation)\n",
 		validationMemory, validationMemory/1000)
 }
 
-// Concurrency and thread safety tests
 func runConcurrencyTests() {
-	fmt.Println("\n🔄 Concurrency Tests")
+	fmt.Println("\nConcurrency Tests")
 	fmt.Println("====================")
 
 	schema := validators.Object(map[string]interface{}{
@@ -238,21 +226,20 @@ func runConcurrencyTests() {
 	}).Required()
 
 	fmt.Println("\n🧵 Testing concurrent validation...")
-	
+
 	start := time.Now()
-	
-	// Run concurrent validations
+
 	done := make(chan bool, 100)
 	for i := 0; i < 100; i++ {
 		go func(id int) {
 			defer func() { done <- true }()
-			
+
 			for j := 0; j < 100; j++ {
 				testData := map[string]interface{}{
 					"id":   id*100 + j,
 					"name": fmt.Sprintf("user_%d_%d", id, j),
 				}
-				
+
 				if err := schema.Validate(testData); err != nil {
 					fmt.Printf("❌ Concurrent validation failed: %v\n", err)
 					return
@@ -261,19 +248,17 @@ func runConcurrencyTests() {
 		}(i)
 	}
 
-	// Wait for all goroutines
 	for i := 0; i < 100; i++ {
 		<-done
 	}
-	
+
 	duration := time.Since(start)
-	fmt.Printf("   ✅ 10,000 concurrent validations completed in %v\n", duration)
-	fmt.Printf("   🚀 Concurrent throughput: %.0f ops/sec\n", 10000.0/duration.Seconds())
+	fmt.Printf("   10,000 concurrent validations completed in %v\n", duration)
+	fmt.Printf("    Concurrent throughput: %.0f ops/sec\n", 10000.0/duration.Seconds())
 }
 
-// Real-world scenario testing
 func runRealWorldScenarios() {
-	fmt.Println("\n🌍 Real-World Scenarios")
+	fmt.Println("\n Real-World SHII")
 	fmt.Println("=======================")
 
 	scenarios := []struct {
@@ -379,36 +364,33 @@ func runRealWorldScenarios() {
 
 	for _, scenario := range scenarios {
 		fmt.Printf("\n🎯 Testing: %s - %s\n", scenario.name, scenario.description)
-		
+
 		schema, testData := scenario.setup()
-		
-		// Benchmark the scenario
+
 		iterations := 1000
 		start := time.Now()
-		
+
 		for i := 0; i < iterations; i++ {
 			if err := schema.Validate(testData); err != nil {
 				fmt.Printf("   ❌ Validation failed: %v\n", err)
 				break
 			}
 		}
-		
+
 		duration := time.Since(start)
-		fmt.Printf("   ⏱️  %d validations: %v (avg: %v)\n", 
+		fmt.Printf("   ⏱️  %d validations: %v (avg: %v)\n",
 			iterations, duration, duration/time.Duration(iterations))
-		fmt.Printf("   🚀 Throughput: %.0f validations/sec\n", 
+		fmt.Printf("   🚀 Throughput: %.0f validations/sec\n",
 			float64(iterations)/duration.Seconds())
 	}
 }
 
-// Comparison tests between different API approaches
 func runComparisonTests() {
 	fmt.Println("\n⚖️  API Comparison Tests")
 	fmt.Println("========================")
 
 	testData := "test_string_123"
 
-	// Test different error key approaches
 	errorKeyApproaches := []struct {
 		name  string
 		setup func() interface{ Validate(interface{}) error }
@@ -450,22 +432,21 @@ func runComparisonTests() {
 	fmt.Println("\n📊 Error Key Approach Performance:")
 	for _, approach := range errorKeyApproaches {
 		schema := approach.setup()
-		
+
 		iterations := 10000
 		start := time.Now()
-		
+
 		for i := 0; i < iterations; i++ {
 			_ = schema.Validate(testData)
 		}
-		
+
 		duration := time.Since(start)
-		fmt.Printf("   %s: %v (%.0f ops/sec)\n", 
+		fmt.Printf("   %s: %v (%.0f ops/sec)\n",
 			approach.name, duration, float64(iterations)/duration.Seconds())
 	}
 
-	// Test state transition performance
 	fmt.Println("\n🔄 State Transition Performance:")
-	
+
 	stateApproaches := []struct {
 		name  string
 		setup func() interface{ Validate(interface{}) error }
@@ -499,53 +480,51 @@ func runComparisonTests() {
 	for _, approach := range stateApproaches {
 		iterations := 10000
 		start := time.Now()
-		
+
 		for i := 0; i < iterations; i++ {
 			schema := approach.setup()
 			_ = schema.Validate(testData)
 		}
-		
+
 		duration := time.Since(start)
-		fmt.Printf("   %s: %v (%.0f ops/sec)\n", 
+		fmt.Printf("   %s: %v (%.0f ops/sec)\n",
 			approach.name, duration, float64(iterations)/duration.Seconds())
 	}
 }
 
-// Generate a comprehensive performance report
 func generatePerformanceReport() {
 	fmt.Println("\n📋 Performance Report Generation")
 	fmt.Println("================================")
 
 	report := struct {
-		Timestamp    string                 `json:"timestamp"`
-		GoVersion    string                 `json:"go_version"`
-		OS           string                 `json:"os"`
-		Arch         string                 `json:"arch"`
-		NumCPU       int                    `json:"num_cpu"`
-		Benchmarks   map[string]interface{} `json:"benchmarks"`
-		Summary      map[string]interface{} `json:"summary"`
+		Timestamp  string                 `json:"timestamp"`
+		GoVersion  string                 `json:"go_version"`
+		OS         string                 `json:"os"`
+		Arch       string                 `json:"arch"`
+		NumCPU     int                    `json:"num_cpu"`
+		Benchmarks map[string]interface{} `json:"benchmarks"`
+		Summary    map[string]interface{} `json:"summary"`
 	}{
-		Timestamp: time.Now().Format(time.RFC3339),
-		GoVersion: runtime.Version(),
-		OS:        runtime.GOOS,
-		Arch:      runtime.GOARCH,
-		NumCPU:    runtime.NumCPU(),
+		Timestamp:  time.Now().Format(time.RFC3339),
+		GoVersion:  runtime.Version(),
+		OS:         runtime.GOOS,
+		Arch:       runtime.GOARCH,
+		NumCPU:     runtime.NumCPU(),
 		Benchmarks: make(map[string]interface{}),
 		Summary: map[string]interface{}{
 			"api_version": "enhanced_dx_v1",
 			"features": []string{
 				"type_safe_state_management",
-				"smart_error_key_autocompletion", 
+				"smart_error_key_autocompletion",
 				"clean_package_interface",
 				"compile_time_safety",
 			},
 		},
 	}
 
-	// Run quick benchmarks for the report
 	quickBenchmarks := []struct {
-		name     string
-		runFunc  func() (time.Duration, int)
+		name    string
+		runFunc func() (time.Duration, int)
 	}{
 		{
 			name: "string_validation",
@@ -579,13 +558,13 @@ func generatePerformanceReport() {
 					"email": validators.Email(),
 					"age":   validators.Number().Integer().Optional(),
 				}).Required()
-				
+
 				testData := map[string]interface{}{
 					"name":  "John",
 					"email": "john@example.com",
 					"age":   30,
 				}
-				
+
 				iterations := 5000
 				start := time.Now()
 				for i := 0; i < iterations; i++ {
@@ -599,14 +578,14 @@ func generatePerformanceReport() {
 	for _, benchmark := range quickBenchmarks {
 		duration, iterations := benchmark.runFunc()
 		opsPerSec := float64(iterations) / duration.Seconds()
-		
+
 		report.Benchmarks[benchmark.name] = map[string]interface{}{
 			"duration_ns":     duration.Nanoseconds(),
 			"iterations":      iterations,
 			"ops_per_second":  opsPerSec,
 			"avg_duration_ns": duration.Nanoseconds() / int64(iterations),
 		}
-		
+
 		fmt.Printf("   ✅ %s: %.0f ops/sec\n", benchmark.name, opsPerSec)
 	}
 
@@ -617,9 +596,9 @@ func generatePerformanceReport() {
 		return
 	}
 
-	filename := fmt.Sprintf("performance_report_%s.json", 
+	filename := fmt.Sprintf("performance_report_%s.json",
 		time.Now().Format("2006-01-02_15-04-05"))
-	
+
 	if err := os.WriteFile(filename, reportJSON, 0644); err != nil {
 		fmt.Printf("❌ Failed to save report: %v\n", err)
 		return
@@ -629,12 +608,10 @@ func generatePerformanceReport() {
 	fmt.Printf("   📊 Report size: %d bytes\n", len(reportJSON))
 }
 
-// Helper function for running Go benchmarks programmatically
 func runGoBenchmark(benchFunc func(*testing.B)) testing.BenchmarkResult {
 	return testing.Benchmark(benchFunc)
 }
 
-// Utility functions for testing
 func warmup(fn func()) {
 	for i := 0; i < 100; i++ {
 		fn()
@@ -651,9 +628,8 @@ func measureMemory(fn func()) uint64 {
 }
 
 func init() {
-	// Set GOMAXPROCS to use all available CPUs
 	runtime.GOMAXPROCS(runtime.NumCPU())
-	
-	fmt.Printf("🖥️  System Info: %s %s, %d CPUs, Go %s\n", 
+
+	fmt.Printf("🖥️  System Info: %s %s, %d CPUs, Go %s\n",
 		runtime.GOOS, runtime.GOARCH, runtime.NumCPU(), runtime.Version())
 }
